refactor: separated widgets
